refactor(tests): Consolidate per-schema spec files
Merge each schema's hand-written, property-based, and edge-case specs
into a single mockX.spec.ts using nested describes (hand-written cases
/ property-based / edge cases). Fold async wrappers and the seven tiny
trivial mockers (Boolean, NaN, Literal, Nullable, NonNullable, Nullish,
Optional) into mockTrivial; absorb the composite edge file into
mockUnion / mockIntersect.
