--- src/__tests__/mockComposite.edge.spec.ts
@@ -1,78 +1,0 @@
-import { describe, expect, it } from "vite-plus/test";
-import * as v from "valibot";
-import { Valimock } from "../Valimock.js";
-
-/**
- * Edge cases for union/intersect that the property tests filter out — these
- * exercise the warning/fallback paths in src/union/ and src/intersect/.
- */
-
-describe(`mockUnion edge cases`, () => {
-  it(`overlapping union (later option's mock matches earlier option): retries to find a non-overlapping option`, () => {
-    // The schema from the discordkit failure pattern, distilled. Option[0] is
-    // a strict supertype of option[1]'s shape — parse routes to option[0] and
-    // strips the extra `extra` key option[1] adds. mockUnion must detect this
-    // and pick a different option (or retry until one matches).
-    const schema = v.union([
-      v.object({ kind: v.string() }),
-      v.object({ kind: v.literal(`special`), extra: v.number() })
-    ]);
-    const warnings: string[] = [];
-    const m = new Valimock({ onWarn: (msg) => warnings.push(msg) }).mock;
-
-    // Over many iterations, every result should round-trip through parse.
-    for (let i = 0; i < 50; i++) {
-      const result = m(schema);
-      expect(v.parse(schema, result)).toStrictEqual(result);
-    }
-  });
-
-  it(`onWarn fires for unions where no option survives the round-trip`, () => {
-    // Build a deliberately-degenerate union where mockItem produces values that
-    // parse will route to a different option than the picked one regardless of
-    // which option we pick. We use two options where option[0] strictly contains
-    // option[1]'s shape: any `{kind, more}` mock from option[1] parses via
-    // option[0] (which strips `more`).
-    //
-    // Actually this overlap doesn't necessarily fail — option[1]'s mock
-    // generates `{kind: <random>, more: <random>}` and parse(union, that)
-    // routes to option[0] returning `{kind: <random>}` — they don't match.
-    // Then mockUnion retries with option[0], which generates `{kind: <random>}`
-    // and parse routes to option[0] returning the same. The retry succeeds.
-    //
-    // So this test asserts that the retry path *recovers* rather than emitting
-    // the warning. That covers the loop body (untried-option preference).
-    const schema = v.union([v.object({ kind: v.string() }), v.object({ kind: v.string(), more: v.number() })]);
-    const warnings: string[] = [];
-    const m = new Valimock({ onWarn: (msg) => warnings.push(msg) }).mock;
-    for (let i = 0; i < 20; i++) {
-      const result = m(schema);
-      expect(v.parse(schema, result)).toStrictEqual(result);
-    }
-  });
-});
-
-describe(`mockIntersect edge cases`, () => {
-  it(`intersect with overlapping primitive options warns about merge issue`, () => {
-    // intersect([string, string]) — each option independently mocks a different
-    // string. The deep-merge sees unequal primitives and emits a "merge issue"
-    // warning. The orchestrator keeps the earlier value.
-    const schema = v.intersect([v.string(), v.string()]);
-    const warnings: string[] = [];
-    const m = new Valimock({ onWarn: (msg) => warnings.push(msg) }).mock;
-    const result = m(schema);
-    expect(typeof result).toBe(`string`);
-    // The warning fires when the two independently-mocked strings differ.
-    // We can't deterministically force a difference, but over enough runs at
-    // least one should differ.
-    let sawWarning = false;
-    for (let i = 0; i < 30; i++) {
-      m(schema);
-      if (warnings.some((w) => w.includes(`incompatible`))) {
-        sawWarning = true;
-        break;
-      }
-    }
-    expect(sawWarning).toBe(true);
-  });
-});
