--- src/__tests__/mockLiteral.spec.ts
@@ -1,16 +1,0 @@
-import { describe, expect, it } from "vite-plus/test";
-import * as v from "valibot";
-import { Valimock } from "../Valimock.js";
-
-const mockSchema = new Valimock().mock;
-
-describe(`mockLiteral`, () => {
-  it.concurrent.each([v.literal(`foo`), v.literal(6)])(
-    `should generate valid mock data (%#)`,
-    { repeats: 5 },
-    (schema) => {
-      const result = mockSchema(schema);
-      expect(v.parse(schema, result)).toStrictEqual(result);
-    }
-  );
-});
