--- src/__tests__/mockNaN.spec.ts
@@ -1,12 +1,0 @@
-import { describe, expect, it } from "vite-plus/test";
-import * as v from "valibot";
-import { Valimock } from "../Valimock.js";
-
-const mockSchema = new Valimock().mock;
-
-describe(`mockNaN`, () => {
-  it.concurrent.each([v.nan()])(`should generate valid mock data (%#)`, { repeats: 5 }, (schema) => {
-    const result = mockSchema(schema);
-    expect(v.parse(schema, result)).toStrictEqual(result);
-  });
-});
