+++ src/__tests__/mockVariant.spec.ts
@@ -1,8 +1,9 @@
+import * as fc from "fast-check";
 import { describe, expect, it } from "vite-plus/test";
 import * as v from "valibot";
 import { Valimock } from "../Valimock.js";
 
-const mockSchema = new Valimock().mock;
+const mockSchema = new Valimock({ onWarn: () => {} }).mock;
 
 enum PetType {
   CAT = `cat`,
@@ -14,12 +15,10 @@ enum PetType {
 const petTypeSchema = v.enum_(PetType);
 
 const petsSchema = v.intersect([
-  // common properties
   v.object({
     name: v.pipe(v.string(), v.nonEmpty()),
     age: v.pipe(v.number(), v.integer())
   }),
-  // type specific properties
   v.variant(`type`, [
     v.object({
       type: v.literal(PetType.CAT),
@@ -95,13 +94,54 @@ const nestedSchema = v.intersect([
   ])
 ]);
 
-describe(`mockVariant`, () => {
-  it.concurrent.each([petsSchema, vehiclesSchema, nestedSchema])(
-    `should generate valid mock data (%#)`,
-    { repeats: 5 },
-    (schema) => {
-      const result = mockSchema(schema);
-      expect(v.parse(schema, result)).toStrictEqual(result);
+const variantSchemaArb: fc.Arbitrary<v.GenericSchema<unknown>> = fc
+  .array(
+    fc
+      .record({
+        tag: fc.string({ minLength: 1, maxLength: 8 }),
+        bodyShape: fc.constantFrom(`string`, `number`, `boolean`)
+      })
+      .filter(({ tag }) => /^[a-zA-Z][a-zA-Z0-9]*$/.test(tag)),
+    { minLength: 2, maxLength: 4 }
+  )
+  .filter((entries) => {
+    // variant requires unique discriminator values across options.
+    const seen = new Set<string>();
+    for (const { tag } of entries) {
+      if (seen.has(tag)) return false;
+      seen.add(tag);
     }
-  );
+    return true;
+  })
+  .map((entries) => {
+    const options = entries.map(({ tag, bodyShape }) => {
+      const body = bodyShape === `string` ? v.string() : bodyShape === `number` ? v.number() : v.boolean();
+      return v.object({ type: v.literal(tag), data: body });
+    });
+    return v.variant(`type`, options as never) as unknown as v.GenericSchema<unknown>;
+  });
+
+describe(`mockVariant`, () => {
+  describe(`hand-written cases`, () => {
+    it.concurrent.each([petsSchema, vehiclesSchema, nestedSchema])(
+      `should generate valid mock data (%#)`,
+      { repeats: 5 },
+      (schema) => {
+        const result = mockSchema(schema);
+        expect(v.parse(schema, result)).toStrictEqual(result);
+      }
+    );
+  });
+
+  describe(`property-based`, () => {
+    it(`every mock value round-trips through Valibot's parse`, () => {
+      fc.assert(
+        fc.property(variantSchemaArb, (schema) => {
+          const result = mockSchema(schema);
+          expect(v.parse(schema, result)).toEqual(result);
+        }),
+        { numRuns: 200 }
+      );
+    });
+  });
 });
